+++ embedded/wheels_esp32/lib/mechnum/mechnum.cpp
@@ -76,7 +76,7 @@ void mechnum::enc_ISR3() {
 }
 
 // Example speed calculation using encoder pulses
-float mechnum::get_speed_enc() {
+int mechnum::get_speed_enc() {
     int pulsesPerRevolution = 150;  // Adjust based on your motor specs
     float speed_encoder = (pulsecount / (float)pulsesPerRevolution);  // Convert to RPM
     pulsecount = 0;  // Reset pulse count for the next cycle
